@@ -1,12 +1,11 @@
 import { ColumnStack, Fit } from "@skbkontur/react-stack-layout";
-import React, { SyntheticEvent } from "react";
+import React from "react";
 
-import { DateTimeRange, DateTimeRangeChange, ICanBeValidated } from "../../Domain/DataTypes/DateTimeRange";
+import { DateTimeRange } from "../../Domain/DataTypes/DateTimeRange";
 import { TimeZone } from "../../Domain/DataTypes/Time";
 import { TimeUtils } from "../../Domain/Utils/TimeUtils";
 
 import { DatePicker } from "./DatePicker";
-import { DateTimePicker } from "./DateTimePicker";
 import styles from "./DateTimeRangePicker.less";
 import { RangeSelector } from "./RangeSelector";
 
@@ -19,7 +18,7 @@ export interface PredefinedRangeDefinition {
 export interface DateTimeRangePickerProps {
     error?: boolean;
     value: DateTimeRange;
-    onChange: DateTimeRangeChange;
+    onChange: (value: DateTimeRange) => void;
     disabled?: boolean;
     timeZone?: TimeZone;
     hideTime?: boolean;
@@ -56,31 +55,19 @@ const defaultPredefinedRanges: PredefinedRangeDefinition[] = [
     },
 ];
 
-export class DateTimeRangePicker extends React.Component<DateTimeRangePickerProps> implements ICanBeValidated {
+export class DateTimeRangePicker extends React.Component<DateTimeRangePickerProps> {
     public static defaultProps = {
         lowerBoundDefaultTime: "00:00",
         upperBoundDefaultTime: "23:59",
         predefinedRanges: defaultPredefinedRanges,
     };
 
-    public fromDatePicker: DatePicker | DateTimePicker | null = null;
-
-    public focus() {
-        if (this.fromDatePicker != null) {
-            this.fromDatePicker.focus();
-        }
-    }
-
-    public handleClickPredefinedRange(e: SyntheticEvent<any>, value: DateTimeRange) {
+    private handleClickPredefinedRange(value: DateTimeRange) {
         const { disabled, onChange } = this.props;
         if (!disabled) {
-            onChange(e, value);
+            onChange(value);
         }
     }
-
-    public setRefToFromDatePicker = (el: DatePicker) => {
-        this.fromDatePicker = el;
-    };
 
     public render(): JSX.Element {
         const {
@@ -89,58 +76,50 @@ export class DateTimeRangePicker extends React.Component<DateTimeRangePickerProp
             error,
             onChange,
             timeZone,
-            hideTime,
             lowerBoundDefaultTime,
             upperBoundDefaultTime,
             predefinedRanges,
             ...rest
         } = this.props;
         const { lowerBound, upperBound } = value;
-        const Picker = hideTime ? DatePicker : DateTimePicker;
-
         const fixedTimezone = TimeUtils.getTimeZoneOffsetOrDefault(timeZone);
 
         return (
             <ColumnStack className={styles.dateRange} {...rest}>
                 <Fit>
                     <span className={styles.dateRangeItem}>
-                        {/* отдельно позже разобрать и пофиксить это место
-                        // @ts-ignore */}
-                        <Picker
+                        <DatePicker
                             data-tid="From"
-                            ref={this.setRefToFromDatePicker}
                             error={error}
                             value={lowerBound}
                             disabled={disabled}
                             defaultTime={lowerBoundDefaultTime}
                             timeZone={fixedTimezone}
-                            onChange={(e: SyntheticEvent<any>, nextValue: Nullable<Date>) =>
-                                onChange(e, { lowerBound: nextValue, upperBound: upperBound })
+                            onChange={(nextValue: Nullable<Date>) =>
+                                onChange({ lowerBound: nextValue, upperBound: upperBound })
                             }
                         />
                     </span>
                     <span className={styles.dateRangeItem}>&mdash;</span>
                     <span className={styles.dateRangeItem}>
-                        {/* отдельно позже разобрать и пофиксить это место
-                        // @ts-ignore */}
-                        <Picker
+                        <DatePicker
                             data-tid="To"
                             error={error}
                             value={upperBound}
                             disabled={disabled}
                             defaultTime={upperBoundDefaultTime}
                             timeZone={fixedTimezone}
-                            onChange={(e: SyntheticEvent<any>, nextValue: Nullable<Date>) =>
-                                onChange(e, { upperBound: nextValue, lowerBound: lowerBound })
+                            onChange={(nextValue: Nullable<Date>) =>
+                                onChange({ upperBound: nextValue, lowerBound: lowerBound })
                             }
                         />
                     </span>
                 </Fit>
-                <Fit className={`${styles.templates} ${hideTime && styles.smallGap} ${disabled && styles.disabled}`}>
+                <Fit className={`${styles.templates} ${styles.smallGap} ${disabled && styles.disabled}`}>
                     {(predefinedRanges || defaultPredefinedRanges).map(x => (
                         <span
                             key={x.tid}
-                            onClick={e => this.handleClickPredefinedRange(e, x.getRange(this.props.timeZone))}
+                            onClick={_ => this.handleClickPredefinedRange(x.getRange(this.props.timeZone))}
                             data-tid={x.tid}>
                             {x.caption}
                         </span>
