@@ -2,9 +2,7 @@ import DeleteIcon from "@skbkontur/react-icons/Delete";
 import ListRowsIcon from "@skbkontur/react-icons/ListRows";
 import RefreshIcon from "@skbkontur/react-icons/Refresh";
 import { ColumnStack, Fill, Fit, RowStack } from "@skbkontur/react-stack-layout";
-import Button from "@skbkontur/react-ui/Button";
-import Link from "@skbkontur/react-ui/Link";
-import Modal from "@skbkontur/react-ui/Modal";
+import { Button, Link, Modal } from "@skbkontur/react-ui";
 import { LocationDescriptor } from "history";
 import React from "react";
 import { Link as RouterLink } from "react-router-dom";
